Modularize template
- Split modules into multiple files to increase readability, maintainability.
